replace method in Replacement now takes String address as argument instead of Page. Added TODO to CrawledLink
